openta/django/pyproject.toml
[tool.poetry]
name = "OpenTA"
version = "0.1.0"
description = ""
authors = ["OpenTA Project <opentaproject@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "3.7.3"
appnope = "0.1.2"
argon2-cffi = "20.1.0"
asgiref = "3.3.1"
async-generator = "1.10"
async-timeout = "3.0.1"
attrs = "20.3.0"
backcall = "0.2.0"
backports-abc = "0.5"
"backports.csv" = "1.0.7"
bleach = "3.2.2"
cachetools = "4.2.0"
certifi = "2020.12.5"
cffi = "1.14.4"
chardet = "3.0.4"
click = "7.1.2"
codetiming = "1.2.0"
dataclasses = "0.6"
decorator = "4.4.2"
deepdiff = "5.2.2"
defusedxml = "0.6.0"
diff-match-patch = "20200713"
Django = "3.1.5"
django-anymail = "8.1"
django-appconf = "1.0.4"
django-autotranslate = "1.1.1"
django-compat = "1.0.15"
django-extensions = "3.1.0"
django-hijack = "2.1.10"
django-imagekit = "4.0.2"
django-import-export = "2.5.0"
django-model-utils = "4.1.1"
django-ratelimit = "3.0.1"
django-rq = "2.4.0"
django-storages = "1.11.1"
django-widget-tweaks = "1.4.8"
djangorestframework = "3.12.2"
entrypoints = "0.3"
et-xmlfile = "1.0.1"
futures = "3.1.1"
google-api-core = "1.25.0"
google-auth = "1.24.0"
google-cloud = "0.34.0"
google-cloud-core = "1.5.0"
google-cloud-translate = "3.0.2"
googleapis-common-protos = "1.52.0"
goslate = "1.5.1"
grpcio = "1.35.0"
httpie = "2.3.0"
httplib2 = "0.18.1"
idna = "2.10"
idna-ssl = "1.1.0"
importlib-metadata = "3.4.0"
IPy = "1.1"
ipykernel = "5.4.3"
ipython = "7.19.0"
ipython-genutils = "0.2.0"
ipywidgets = "7.6.3"
jdcal = "1.4.1"
jedi = "0.18.0"
Jinja2 = "2.11.2"
jsonschema = "3.2.0"
jupyter-client = "6.1.11"
jupyter-core = "4.7.0"
jupyterlab-pygments = "0.1.2"
jupyterlab-widgets = "1.0.0"
kubernetes = "12.0.1"
libcst = "0.3.16"
lxml = "4.6.2"
MarkupPy = "1.14"
MarkupSafe = "1.1.1"
mistune = "0.8.4"
mpmath = "1.1.0"
multidict = "5.1.0"
mypy-extensions = "0.4.3"
nbclient = "0.5.1"
nbconvert = "6.0.7"
nbformat = "5.1.2"
nest-asyncio = "1.4.3"
notebook = "6.2.0"
numpy = "1.19.5"
oauth2 = "1.9.0.post1"
oauthlib = "3.1.0"
odfpy = "1.4.1"
openpyxl = "3.0.6"
ordered-set = "4.0.2"
packaging = "20.8"
pandocfilters = "1.4.3"
parso = "0.8.1"
pexpect = "4.8.0"
pickleshare = "0.7.5"
pilkit = "2.0"
Pillow = "8.1.0"
polib = "1.1.0"
portalocker = "2.1.0"
prometheus-client = "0.9.0"
prompt-toolkit = "3.0.14"
proto-plus = "1.13.0"
protobuf = "3.14.0"
psycopg2-binary = "2.8.6"
ptyprocess = "0.7.0"
pyasn1 = "0.4.8"
pyasn1-modules = "0.2.8"
pycparser = "2.20"
pydotplus = "2.0.2"
Pygments = "2.7.4"
pygraphviz = "1.6"
PyLTI = "0.7.0"
pyparsing = "2.4.7"
PyPDF2 = "1.26.0"
pyrsistent = "0.17.3"
PySocks = "1.7.1"
python-dateutil = "2.8.1"
pytz = "2020.5"
PyYAML = "5.4.1"
pyzmq = "21.0.1"
redis = "3.5.3"
requests = "2.25.1"
requests-oauthlib = "1.3.0"
requests-toolbelt = "0.9.1"
rq = "1.7.0"
rsa = "4.7"
selenium = "3.141.0"
Send2Trash = "1.5.0"
simplegeneric = "0.8.1"
six = "1.15.0"
sqlparse = "0.4.1"
sympy = "1.7.1"
tablib = "3.0.0"
tabulate = "0.8.7"
termcolor = "1.1.0"
terminado = "0.9.2"
testpath = "0.4.4"
tornado = "6.1"
traitlets = "5.0.5"
typing-inspect = "0.6.0"
uritemplate = "3.0.1"
urllib3 = "1.26.2"
wcwidth = "0.2.5"
webencodings = "0.5.1"
websocket-client = "0.57.0"
widgetsnbextension = "3.5.1"
wolframclient = "1.1.4"
xlrd = "2.0.1"
XlsxWriter = "1.3.7"
xlwt = "1.3.0"
xmljson = "0.2.1"
xmltodict = "0.12.0"
python-slugify = "4.0.1"
yarl = "1.6.3"
zipp = "3.4.0"
django-memcached = "0.1.2"
memcache = "0.2.0"
python-memcached = "1.59"
django-redis-cache = "3.0.0"
django-redis = "5.0.0"
pylibmc = "1.6.1"
lti = "0.9.5"
pandas = "1.3.2"
djangorestframework-dataclasses = "^1.0.0"
debugpy = "^1.5.1"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "3.2"
selenium = "3.141.0"
pytest-django = "^4.4.0"
coverage = "^6.0.2"
Faker = "^9.6.0"
pytest-mock = "^3.6.1"
django-stubs = "^1.9.0"
djangorestframework-stubs = "^1.4.0"
flake8 = "^4.0.1"
pytest-cov = "^3.0.0"
google-api-python-client-stubs = "^1.7.1"
lxml-stubs = "^0.2.0"
tox = "^3.24.4"

[tool.isort]
profile = "black"

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
exclude="migrations"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
